requirement.txt
faiss-cpu
numpy
tqdm
transformers
openai
torch
pandas
